run plug in prior to processing css modules
